+++ firebase-app-distribution/src/main/java/com/google/firebase/appdistribution/FirebaseAppDistribution.java
@@ -15,6 +15,8 @@
 package com.google.firebase.appdistribution;
 
 import static com.google.firebase.appdistribution.FirebaseAppDistributionException.Status.AUTHENTICATION_FAILURE;
+import static com.google.firebase.appdistribution.TaskUtils.safeSetTaskException;
+import static com.google.firebase.appdistribution.TaskUtils.safeSetTaskResult;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -28,6 +30,8 @@ import com.google.android.gms.common.internal.Preconditions;
 import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.appdistribution.Constants.ErrorMessages;
+import com.google.firebase.appdistribution.FirebaseAppDistributionException.Status;
 import com.google.firebase.appdistribution.internal.AppDistributionReleaseInternal;
 import com.google.firebase.installations.FirebaseInstallationsApi;
 import org.jetbrains.annotations.Nullable;
@@ -48,6 +52,8 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
   private Task<AppDistributionRelease> cachedCheckForNewReleaseTask;
 
   private AppDistributionReleaseInternal cachedNewRelease;
+  private AlertDialog updateDialog;
+  private boolean updateDialogShown;
   private final SignInStorage signInStorage;
 
   /** Constructor for FirebaseAppDistribution */
@@ -242,20 +248,15 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
   @Override
   public void onActivityStarted(@NonNull Activity activity) {
     LogWrapper.getInstance().d("Started activity: " + activity.getClass().getName());
-  }
+    // SignInResultActivity and InstallActivity are internal to the SDK and should not be treated as
+    // reentering the app
+    if (activity instanceof SignInResultActivity || activity instanceof InstallActivity) {
+      return;
+    }
 
-  @Override
-  public void onActivityResumed(@NonNull Activity activity) {
-    LogWrapper.getInstance().d("Resumed activity: " + activity.getClass().getName());
     // If app resumes and aab update task is in progress, assume that installation didn't happen so
     // cancel the task
     updateAppClient.tryCancelAabUpdateTask();
-
-    // SignInResultActivity is only opened after successful redirection from signIn flow,
-    // should not be treated as reentering the app
-    if (activity instanceof SignInResultActivity) {
-      return;
-    }
 
     // Throw error if app reentered during sign in
     if (this.testerSignInClient.isCurrentlySigningIn()) {
@@ -269,13 +270,36 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
   }
 
   @Override
+  public void onActivityResumed(@NonNull Activity activity) {
+    LogWrapper.getInstance().d("Resumed activity: " + activity.getClass().getName());
+
+    if (activity instanceof SignInResultActivity || activity instanceof InstallActivity) {
+      return;
+    }
+
+    this.currentActivity = activity;
+    this.updateAppClient.setCurrentActivity(activity);
+    this.testerSignInClient.setCurrentActivity(activity);
+  }
+
+  @Override
   public void onActivityPaused(@NonNull Activity activity) {
     LogWrapper.getInstance().d("Paused activity: " + activity.getClass().getName());
+    if (this.currentActivity == activity) {
+      this.currentActivity = null;
+      this.updateAppClient.setCurrentActivity(null);
+      this.testerSignInClient.setCurrentActivity(null);
+    }
   }
 
   @Override
   public void onActivityStopped(@NonNull Activity activity) {
     LogWrapper.getInstance().d("Stopped activity: " + activity.getClass().getName());
+    if (this.currentActivity == activity) {
+      this.currentActivity = null;
+      this.updateAppClient.setCurrentActivity(null);
+      this.testerSignInClient.setCurrentActivity(null);
+    }
   }
 
   @Override
@@ -286,10 +310,22 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
   @Override
   public void onActivityDestroyed(@NonNull Activity activity) {
     LogWrapper.getInstance().d("Destroyed activity: " + activity.getClass().getName());
+    if (updateDialogShown) {
+      setCachedUpdateIfNewReleaseCompletionError(
+          new FirebaseAppDistributionException(
+              ErrorMessages.UPDATE_CANCELED, Status.INSTALLATION_CANCELED));
+    }
+
     if (this.currentActivity == activity) {
       this.currentActivity = null;
       this.updateAppClient.setCurrentActivity(null);
       this.testerSignInClient.setCurrentActivity(null);
+    }
+
+    if (activity instanceof InstallActivity) {
+      // Since install activity is destroyed but app is still active, installation has failed /
+      // cancelled.
+      updateAppClient.trySetInstallTaskError();
     }
   }
 
@@ -305,8 +341,8 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
 
   private UpdateTaskImpl showUpdateAlertDialog(AppDistributionRelease newRelease) {
     Context context = firebaseApp.getApplicationContext();
-    AlertDialog alertDialog = new AlertDialog.Builder(currentActivity).create();
-    alertDialog.setTitle(context.getString(R.string.update_dialog_title));
+    updateDialog = new AlertDialog.Builder(currentActivity).create();
+    updateDialog.setTitle(context.getString(R.string.update_dialog_title));
 
     StringBuilder message =
         new StringBuilder(
@@ -318,8 +354,8 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
       message.append(String.format("\n\nRelease notes: %s", newRelease.getReleaseNotes()));
     }
 
-    alertDialog.setMessage(message);
-    alertDialog.setButton(
+    updateDialog.setMessage(message);
+    updateDialog.setButton(
         AlertDialog.BUTTON_POSITIVE,
         context.getString(R.string.update_yes_button),
         (dialogInterface, i) -> {
@@ -333,7 +369,7 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
           }
         });
 
-    alertDialog.setButton(
+    updateDialog.setButton(
         AlertDialog.BUTTON_NEGATIVE,
         context.getString(R.string.update_no_button),
         (dialogInterface, i) -> {
@@ -347,26 +383,21 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
                     .build());
             setCachedUpdateIfNewReleaseCompletionError(
                 new FirebaseAppDistributionException(
-                    Constants.ErrorMessages.UPDATE_CANCELED,
-                    FirebaseAppDistributionException.Status.INSTALLATION_CANCELED));
+                    ErrorMessages.UPDATE_CANCELED, Status.INSTALLATION_CANCELED));
           }
         });
 
-    alertDialog.show();
+    updateDialog.show();
+    updateDialogShown = true;
     synchronized (updateTaskLock) {
       return cachedUpdateIfNewReleaseTask;
     }
   }
 
-  void setInstallationResult(int resultCode) {
-    this.updateAppClient.setInstallationResult(resultCode);
-  }
-
   private void setCachedUpdateIfNewReleaseCompletionError(FirebaseAppDistributionException e) {
     synchronized (updateTaskLock) {
-      if (cachedUpdateIfNewReleaseTask != null && !cachedUpdateIfNewReleaseTask.isComplete()) {
-        cachedUpdateIfNewReleaseTask.setException(e);
-      }
+      safeSetTaskException(cachedUpdateIfNewReleaseTask, e);
+      dismissUpdateDialog();
     }
   }
 
@@ -387,7 +418,15 @@ public class FirebaseAppDistribution implements Application.ActivityLifecycleCal
 
   private void setCachedUpdateIfNewReleaseResult() {
     synchronized (updateTaskLock) {
-      cachedUpdateIfNewReleaseTask.setResult();
+      safeSetTaskResult(cachedUpdateIfNewReleaseTask);
+      dismissUpdateDialog();
+    }
+  }
+
+  private void dismissUpdateDialog() {
+    if (updateDialog != null) {
+      updateDialog.dismiss();
+      updateDialogShown = false;
     }
   }
 }