+++ firebase-firestore/src/main/java/com/google/firebase/firestore/remote/FirestoreCallCredentials.java
@@ -14,8 +14,11 @@
 
 package com.google.firebase.firestore.remote;
 
+import com.google.android.gms.tasks.Task;
+import com.google.android.gms.tasks.Tasks;
 import com.google.firebase.FirebaseApiNotAvailableException;
 import com.google.firebase.firestore.auth.CredentialsProvider;
+import com.google.firebase.firestore.auth.User;
 import com.google.firebase.firestore.util.Logger;
 import com.google.firebase.internal.api.FirebaseNoSignedInUserException;
 import io.grpc.CallCredentials;
@@ -31,10 +34,16 @@ final class FirestoreCallCredentials extends CallCredentials {
   private static final Metadata.Key<String> AUTHORIZATION_HEADER =
       Metadata.Key.of("Authorization", Metadata.ASCII_STRING_MARSHALLER);
 
-  private final CredentialsProvider credentialsProvider;
+  private static final Metadata.Key<String> X_FIREBASE_APPCHECK =
+      Metadata.Key.of("x-firebase-appcheck", Metadata.ASCII_STRING_MARSHALLER);
 
-  FirestoreCallCredentials(CredentialsProvider provider) {
-    credentialsProvider = provider;
+  private final CredentialsProvider<User> authProvider;
+  private final CredentialsProvider<String> appCheckProvider;
+
+  FirestoreCallCredentials(
+      CredentialsProvider<User> authProvider, CredentialsProvider<String> appCheckProvider) {
+    this.authProvider = authProvider;
+    this.appCheckProvider = appCheckProvider;
   }
 
   // Don't have @Override to avoid breaking build when method is removed from interface
@@ -44,31 +53,53 @@ final class FirestoreCallCredentials extends CallCredentials {
   @Override
   public void applyRequestMetadata(
       RequestInfo requestInfo, Executor executor, final MetadataApplier metadataApplier) {
-    credentialsProvider
-        .getToken()
-        .addOnSuccessListener(
+    Task<String> authTask = authProvider.getToken();
+    Task<String> appCheckTask = appCheckProvider.getToken();
+
+    Tasks.whenAll(authTask, appCheckTask)
+        .addOnCompleteListener(
             executor,
-            token -> {
-              Logger.debug(LOG_TAG, "Successfully fetched token.");
+            unused -> {
               Metadata metadata = new Metadata();
-              if (token != null) {
-                metadata.put(AUTHORIZATION_HEADER, "Bearer " + token);
-              }
-              metadataApplier.apply(metadata);
-            })
-        .addOnFailureListener(
-            executor,
-            exception -> {
-              if (exception instanceof FirebaseApiNotAvailableException) {
-                Logger.debug(LOG_TAG, "Firebase Auth API not available, not using authentication.");
-                metadataApplier.apply(new Metadata());
-              } else if (exception instanceof FirebaseNoSignedInUserException) {
-                Logger.debug(LOG_TAG, "No user signed in, not using authentication.");
-                metadataApplier.apply(new Metadata());
+
+              if (authTask.isSuccessful()) {
+                String token = authTask.getResult();
+                Logger.debug(LOG_TAG, "Successfully fetched auth token.");
+                if (token != null) {
+                  metadata.put(AUTHORIZATION_HEADER, "Bearer " + token);
+                }
               } else {
-                Logger.warn(LOG_TAG, "Failed to get token: %s.", exception);
-                metadataApplier.fail(Status.UNAUTHENTICATED.withCause(exception));
+                Exception exception = authTask.getException();
+                if (exception instanceof FirebaseApiNotAvailableException) {
+                  Logger.debug(
+                      LOG_TAG, "Firebase Auth API not available, not using authentication.");
+                } else if (exception instanceof FirebaseNoSignedInUserException) {
+                  Logger.debug(LOG_TAG, "No user signed in, not using authentication.");
+                } else {
+                  Logger.warn(LOG_TAG, "Failed to get auth token: %s.", exception);
+                  metadataApplier.fail(Status.UNAUTHENTICATED.withCause(exception));
+                  return;
+                }
               }
+
+              if (appCheckTask.isSuccessful()) {
+                String token = appCheckTask.getResult();
+                if (token != null && !token.isEmpty()) {
+                  Logger.debug(LOG_TAG, "Successfully fetched AppCheck token.");
+                  metadata.put(X_FIREBASE_APPCHECK, token);
+                }
+              } else {
+                Exception exception = appCheckTask.getException();
+                if (exception instanceof FirebaseApiNotAvailableException) {
+                  Logger.debug(LOG_TAG, "Firebase AppCheck API not available.");
+                } else {
+                  Logger.warn(LOG_TAG, "Failed to get AppCheck token: %s.", exception);
+                  metadataApplier.fail(Status.UNAUTHENTICATED.withCause(exception));
+                  return;
+                }
+              }
+
+              metadataApplier.apply(metadata);
             });
   }
 }