+++ firebase-perf/src/test/java/com/google/firebase/perf/metrics/AppStartTraceTest.java
@@ -18,6 +18,7 @@ import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 import android.app.Activity;
@@ -33,7 +34,9 @@ import com.google.firebase.perf.util.Constants;
 import com.google.firebase.perf.util.Timer;
 import com.google.firebase.perf.v1.ApplicationProcessState;
 import com.google.firebase.perf.v1.TraceMetric;
+import com.google.testing.timing.FakeScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -84,10 +87,17 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     appStartHRT = FirebasePerfProvider.getAppStartTime().getHighResTime();
   }
 
+  @After
+  public void reset() {
+    SessionManager.getInstance()
+        .setPerfSession(com.google.firebase.perf.session.PerfSession.create());
+  }
+
   /** Test activity sequentially goes through onCreate()->onStart()->onResume() state change. */
   @Test
   public void testLaunchActivity() {
-    AppStartTrace trace = new AppStartTrace(transportManager, clock);
+    FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
+    AppStartTrace trace = new AppStartTrace(transportManager, clock, fakeExecutorService);
     // first activity goes through onCreate()->onStart()->onResume() state change.
     currentTime = 1;
     trace.onActivityCreated(activity1, bundle);
@@ -95,7 +105,9 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     trace.onActivityStarted(activity1);
     currentTime = 3;
     trace.onActivityResumed(activity1);
+    fakeExecutorService.runAll();
     verifyFinalState(activity1, trace, 1, 2, 3);
+
     // same activity goes through onCreate()->onStart()->onResume() state change again.
     // should have no effect on AppStartTrace.
     currentTime = 4;
@@ -104,6 +116,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     trace.onActivityStarted(activity1);
     currentTime = 6;
     trace.onActivityResumed(activity1);
+    fakeExecutorService.runAll();
     verifyFinalState(activity1, trace, 1, 2, 3);
 
     // a different activity goes through onCreate()->onStart()->onResume() state change.
@@ -114,6 +127,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     trace.onActivityStarted(activity2);
     currentTime = 9;
     trace.onActivityResumed(activity2);
+    fakeExecutorService.runAll();
     verifyFinalState(activity1, trace, 1, 2, 3);
   }
 
@@ -158,7 +172,8 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
    */
   @Test
   public void testInterleavedActivity() {
-    AppStartTrace trace = new AppStartTrace(transportManager, clock);
+    FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
+    AppStartTrace trace = new AppStartTrace(transportManager, clock, fakeExecutorService);
     // first activity onCreate()
     currentTime = 1;
     trace.onActivityCreated(activity1, bundle);
@@ -178,6 +193,7 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     trace.onActivityResumed(activity2);
     Assert.assertEquals(activity1, trace.getLaunchActivity());
     Assert.assertEquals(activity2, trace.getAppStartActivity());
+    fakeExecutorService.runAll();
     verifyFinalState(activity2, trace, 1, 3, 4);
 
     // first activity continues.
@@ -185,12 +201,14 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
     trace.onActivityStarted(activity1);
     currentTime = 6;
     trace.onActivityResumed(activity1);
+    fakeExecutorService.runAll();
     verifyFinalState(activity2, trace, 1, 3, 4);
   }
 
   @Test
   public void testDelayedAppStart() {
-    AppStartTrace trace = new AppStartTrace(transportManager, clock);
+    FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
+    AppStartTrace trace = new AppStartTrace(transportManager, clock, fakeExecutorService);
     // Delays activity creation after 1 minute from app start time.
     currentTime = appStartTime + TimeUnit.MINUTES.toMicros(1) + 1;
     trace.onActivityCreated(activity1, bundle);
@@ -210,7 +228,8 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
 
   @Test
   public void testStartFromBackground() {
-    AppStartTrace trace = new AppStartTrace(transportManager, clock);
+    FakeScheduledExecutorService fakeExecutorService = new FakeScheduledExecutorService();
+    AppStartTrace trace = new AppStartTrace(transportManager, clock, fakeExecutorService);
     trace.setIsStartFromBackground();
     trace.onActivityCreated(activity1, bundle);
     Assert.assertNull(trace.getOnCreateTime());
@@ -228,14 +247,20 @@ public class AppStartTraceTest extends FirebasePerformanceTestBase {
   }
 
   @Test
-  public void testFirebasePerfProviderOnAttachInfoUpdatesPerfSession() {
+  public void testFirebasePerfProviderOnAttachInfo_initializesGaugeCollection() {
+    com.google.firebase.perf.session.PerfSession mockPerfSession =
+        mock(com.google.firebase.perf.session.PerfSession.class);
+    when(mockPerfSession.sessionId()).thenReturn("sessionId");
+    when(mockPerfSession.isGaugeAndEventCollectionEnabled()).thenReturn(true);
+
+    SessionManager.getInstance().setPerfSession(mockPerfSession);
     String oldSessionId = SessionManager.getInstance().perfSession().sessionId();
-    Assert.assertNotNull(oldSessionId);
     Assert.assertEquals(oldSessionId, SessionManager.getInstance().perfSession().sessionId());
 
     FirebasePerfProvider provider = new FirebasePerfProvider();
     provider.attachInfo(ApplicationProvider.getApplicationContext(), new ProviderInfo());
 
-    Assert.assertNotEquals(oldSessionId, SessionManager.getInstance().perfSession().sessionId());
+    Assert.assertEquals(oldSessionId, SessionManager.getInstance().perfSession().sessionId());
+    verify(mockPerfSession, times(2)).isGaugeAndEventCollectionEnabled();
   }
 }