+++ firebase-firestore/CHANGELOG.md
@@ -1,3 +1,10 @@
+Android changes are not released automatically. Ensure that changes are released 
+by opting into a release at 
+[go/firebase-android-release](http:go/firebase-android-release) (Googlers only).
+
+# Unreleased
+- [feature] Added support for Firebase AppCheck.
+
 # 23.0.4
 - [fixed] Fixed an issue where some fields were missed when copying in the
   `FirebaseFirestoreSettings.Builder` copy constructor.